Connected Database

+++ src/main/java/com/example/TestProject/service/InventoryService.java
@@ -4,5 +4,5 @@ import com.example.TestProject.model.Inventory;
 
 public interface InventoryService {
     Inventory saveDevice(Inventory inventory);
-    Inventory getInventory(Integer id);
+    Inventory getInventory(Long id);
 }
